Moved languages to shared.
